requirements.txt
fastapi==0.89.1
uvicorn[standard]==0.20.0
psycopg2-binary==2.9.5
h3==3.7.6
PyJWT==2.6.0